Added a shorthand for custom grids like `2 of 5`
This feature is a carryover from Neat 1.x. While it should be considered an
advanced feature, It can be useful on a component level. The "official" way to
do this is to create a map for your sub-grid and pass it in to the mixin.

To achieve this, `neat-column-ratio` determines the ratio of the `$column`
to the total columns as defined in `$neat-grid` if it detects multiple values (a
list), it will pass the value to the `neat-parse-column`. This function will
return the third value assuming this is meant to be the total column value.