Refactor `neat-column-width` to check if gutter is zero
While `span-columns` contained a check for when the gutter equals zero.
Since calc operations fail when you attempt to add or subtract zero, an
operation like `calc(100% - 0)` would need to be output as `calc(100%)`.
This opperation has been moved to the `neat-column-width` function so it
can be used by the other mixins.