Remove unitless comparison
This addresses a change to Sass that raises this deprecation warning in
apps running Neat:

DEPRECATION WARNING on line 19 of
…/core/neat/functions/_neat-column-width.scss:
The result of `0px == 0` will be `false` in future releases of Sass.
Unitless numbers will no longer be equal to the same numbers with units.

More info at https://github.com/sass/sass/issues/1496